requirements.txt
python-telegram-bot==20.0
httpx~=0.23.1
python-dotenv==0.21.0
cachetools==5.2.0
ratelimit==2.2.1
base58==2.1.0
